QCC/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "qcc"
version = "0.1.0"
description = "Quantum Cellular Computing framework"
authors = ["QCC Team <info@qcc-project.org>"]
readme = "README.md"
repository = "https://github.com/ovladon/QCC"
documentation = "https://qcc-project.org/docs"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
pydantic = "^2.0.0"
websockets = "^11.0.0"
aiohttp = "^3.8.4"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
httpx = "^0.24.0"
cryptography = "^40.0.0"
pyjwt = "^2.6.0"
bcrypt = "^4.0.0"
passlib = "^1.7.4"
qiskit = "^0.42.0"
pyopenssl = "^23.0.0"
web3 = "^6.0.0"
eth-typing = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^6.2.0"
sphinx-rtd-theme = "^1.2.0"

[tool.poetry.scripts]
qcc = "qcc.cli:main"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
